master

+++ commands/csgo.js
@@ -1,6 +1,7 @@
 const fetch = require('node-fetch');
 const index = require('../index');
 const rtt = require('../methods/roundToTwo');
+const writelog = require('../methods/writelog');
 
 module.exports = async (msg) => {
     const member = msg.member;
@@ -9,14 +10,19 @@ module.exports = async (msg) => {
 
     let sid, kills, deths, plantedbombs, defusedbombs, hostages, knifekills, headshots, shotsfired, shotshit, mvps, kd, headshotp;
 
+    if(channel.id !== index.channelid) return;
+
     if(!index.file[id]) {
         msg.delete();
-        return channel.reply(`Du hast keine SteamID zu deinen Account gebunden! Bitte benutze "!steamid"`);
+        return channel.send(`Du hast keine SteamID zu deinen Account gebunden! Bitte benutze "!steamid"`);
     }
     sid = index.file[id].sid;
 
     const cs = await fetch(`https://api.steampowered.com/ISteamUserStats/GetUserStatsForGame/v2/?key=${index.steamkey}&steamid=${sid}&appid=730`).then(res => res.json());
-    if (cs === undefined) return channel.send(`Konnte keine Daten für ${member.user.username} finden! Eventuell ist das Profil Privat, oder Steam ist Down!`);
+    if (cs === undefined) {
+        writelog(`Konnte keine Daten für ${member.user.username} finden! Profil Privat? Steam Down?`);
+        return channel.send(`Konnte keine Daten für ${member.user.username} finden! Eventuell ist das Profil Privat, oder Steam ist Down!`);
+    }
     cs.playerstats.stats.forEach(element => {
         if (element.name === "total_kills") kills = element.value;
         if (element.name === "total_deaths") deaths = element.value;
