+++ commands/faceit.js
@@ -1,6 +1,7 @@
 const fetch = require('node-fetch');
 const index = require('../index');
 const csgo = require('./csgo');
+const writelog = require('../methods/writelog');
 
 module.exports = async (msg) => {
     const member = msg.member;
@@ -9,9 +10,11 @@ module.exports = async (msg) => {
 
     let fic, sid, m, w, cws, kd, wr, lws, ah, sl, elo;
 
+    if(channel.id !== index.channelid) return;
+
     if (!index.file[id]) {
         msg.delete();
-        return channel.reply(`Du hast keine SteamID zu deinen Account gebunden! Bitte benutze "!steamid"`);
+        return channel.send(`Du hast keine SteamID zu deinen Account gebunden! Bitte benutze "!steamid"`);
     }
     fic = index.file[id].fic;
     sid = index.file[id].sid;
@@ -22,12 +25,18 @@ module.exports = async (msg) => {
     }
 
     const f1 = await fetch(`https://open.faceit.com/data/v4/players/${fic}/stats/csgo`, { method: 'GET', headers: header }).then(res => res.json());
-    if (f1 === undefined) return channel.send(`Konnte keine Daten für ${member.user.username} finden! Eventuell ist die FaceIT API Down`);
+    if (f1 === undefined) {
+        writelog(`Konnte keine Daten für ${member.user.username} finden! FaceIT API Down?`);
+        return channel.send(`Konnte keine Daten für ${member.user.username} finden! Eventuell ist die FaceIT API Down`);
+    }
     m = f1.lifetime["Matches"], w = f1.lifetime["Wins"], cws = f1.lifetime["Current Win Streak"], kd = f1.lifetime["Average K/D Ratio"], wr = f1.lifetime["Win Rate %"], lws = f1.lifetime["Longest Win Streak"], ah = f1.lifetime["Average Headshots %"];
 
 
     const f2 = await fetch(`https://open.faceit.com/data/v4/players?game=csgo&game_player_id=${sid}`, { method: 'GET', headers: header }).then(res => res.json());
-    if (f2 === undefined) return channel.send(`Konnte keine Daten für ${member.user.username} finden! Eventuell ist die FaceIT API Down`);
+    if (f2 === undefined) {
+        writelog(`Konnte keine Daten für ${member.user.username} finden! FaceIT API Down?`);
+        return channel.send(`Konnte keine Daten für ${member.user.username} finden! Eventuell ist die FaceIT API Down`);
+    }
     sl = f2.games.csgo["skill_level"], elo = f2.games.csgo["faceit_elo"];
 
     let mbd = new index.discord.MessageEmbed()
